+++ concordium-sdk/src/main/java/com/concordium/sdk/transactions/TransactionFactory.java
@@ -69,6 +69,26 @@ public class TransactionFactory {
 
 
     /**
+     * Creates a new {@link TransferScheduleTransaction.TransferScheduleTransactionBuilder} for
+     * creating a {@link TransferScheduleTransaction}
+     *
+     * @return the builder for a {@link TransferScheduleTransaction}
+     */
+    public static TransferScheduleTransaction.TransferScheduleTransactionBuilder newScheduledTransfer() {
+        return TransferScheduleTransaction.builder();
+    }
+
+    /**
+     * Creates a new {@link TransferScheduleWithMemoTransaction.TransferScheduleWithMemoTransactionBuilder} for
+     * creating a {@link TransferScheduleWithMemoTransaction}
+     *
+     * @return the builder for a {@link TransferScheduleWithMemoTransaction}
+     */
+    public static TransferScheduleWithMemoTransaction.TransferScheduleWithMemoTransactionBuilder newScheduledTransferWithMemo() {
+        return TransferScheduleWithMemoTransaction.builder();
+    }
+
+    /**
      * Creates a new {@link UpdateCredentialKeysTransaction.UpdateCredentialKeysTransactionBuilder} for
      * creating a {@link UpdateCredentialKeysTransaction}
      *