+++ concordium-sdk/src/test/java/com/concordium/sdk/transactions/TransactionFactoryTest.java
@@ -45,6 +45,18 @@ public class TransactionFactoryTest {
     }
 
     @Test
+    public  void testCanTransferScheduleFactory() {
+        val builder = TransactionFactory.newScheduledTransfer();
+        assertNotNull(builder);
+    }
+
+    @Test
+    public  void testCanTransferSchedulWithMemoFactory() {
+        val builder = TransactionFactory.newScheduledTransferWithMemo();
+        assertNotNull(builder);
+    }
+
+    @Test
     public void testCanCreateUpdateCredentialKeysFactory() {
         val builder = TransactionFactory.newUpdateCredentialKeys();
         assertNotNull(builder);