+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Unreleased changes
+- Added support for Transfer With Schedule.
+- Added support for Transfer With Schedule With Memo.
 
 ## 4.0.0
 - Support for Protocol 5.